libidn (1.41-1) unstable; urgency=medium
.
  * [427112b] Simplify upstream self-check invocation, and more reporting.
  * [007637e] Fix lintian override.
  * [8be0e91] New upstream version 1.41
  * [cf23485] Drop no longer needed lintian override.
.
libidn (1.40-1) unstable; urgency=medium
.
  * [b15bd7e] Drop useless override_dh_auto_install.
  * [7432a06] Use Standards-Version 4.6.1.
  * [af2de03] Use ftp.gnu.org in watchfile.
  * [e1959da] New upstream version 1.40
  * [199f63e] Drop source-contains-prebuilt-java-object lintian override.
.
libidn (1.38-4) unstable; urgency=medium
.
  * [be4e6bb] Use Standards-Version 4.6.0.
.
libidn (1.38-3) unstable; urgency=medium
.
  * Upload to unstable.
.
libidn (1.38-2) experimental; urgency=medium
.
  * [f8d4c77] Fix autopkgtest on hosts with non-UTF8 locale.
  * [6e1f37a] Use gbp dch id-length 7.
.
libidn (1.38-1) experimental; urgency=medium
.
  * New upstream release.  Closes: #966408.
    - All our patches have been fixed upstream, removing 0001-0005.
  * Add d/source/options to extend-diff-ignore javadoc ZIP's.
  * Add file FAQ and remove file TODO from docs.
  * Don't ship libidn.pdf, we never rebuilt it from source,
    and upstream has stopped shipping it.
  * Drop libidn-java (<10 popcon, complicates build/packaging and cross).
  * libidn12 replaces libidn11.
  * libidn-dev replaces libidn11-dev, but left as transitional.
  * Update d/copyright for path changes.
  * Make source-contains-autogenerated-gperf-data override more specific.
  * Library libidn.so is back in /usr/lib (was /lib).
  * Ignore intentional source-contains-prebuilt-javascript-object
    and very-long-line-length-in-source-file.
  * Drop emacs interface, no longer working with modern Emacs.
  * Move libidn.info from -dev to idn package (referenced by idn.1).
  * Make idn autopkgtest test the core of the functionality a bit more.
.
  [ Ondřej Nový <onovy@debian.org> ]
  * d/changelog: Remove trailing whitespaces
.
  [ John Scott ]
  * d/copyright: Revise by consolidating files and license paragraphs.
  * Add good minimal upstream signing key.
  * d/watch: Upgrade to version 4 and use GNU's mirrors.

+++ examples/example3.c
@@ -1,5 +1,5 @@
 /* example3.c --- Example ToASCII() code showing how to use Libidn.
- * Copyright (C) 2002-2016 Simon Josefsson
+ * Copyright (C) 2002-2022 Simon Josefsson
  *
  * This file is part of GNU Libidn.
  *
@@ -14,7 +14,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  */
 
@@ -56,7 +56,7 @@ main (void)
 
   printf ("Read string (length %ld): ", (long int) strlen (buf));
   for (i = 0; i < strlen (buf); i++)
-    printf ("%02x ", buf[i] & 0xFF);
+    printf ("%02x ", (unsigned) buf[i] & 0xFF);
   printf ("\n");
 
   rc = idna_to_ascii_lz (buf, &p, 0);
@@ -68,7 +68,7 @@ main (void)
 
   printf ("ACE label (length %ld): '%s'\n", (long int) strlen (p), p);
   for (i = 0; i < strlen (p); i++)
-    printf ("%02x ", p[i] & 0xFF);
+    printf ("%02x ", (unsigned) p[i] & 0xFF);
   printf ("\n");
 
   free (p);
