+++ tests/tst_pr29.c
@@ -1,5 +1,5 @@
 /* tst_pr29.c --- Self tests for pr29_*().
- * Copyright (C) 2004-2016 Simon Josefsson
+ * Copyright (C) 2004-2022 Simon Josefsson
  *
  * This file is part of GNU Libidn.
  *
@@ -14,7 +14,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  */
 
@@ -81,17 +81,17 @@ static const struct tv tv[] = {
    {0x1100, 0x00AA, 0x1161, 0},
    PR29_SUCCESS},
   {
-    /* http://lists.gnu.org/archive/html/help-libidn/2012-01/msg00008.html */
-    "Infloop",
-    3,
-    {0x1100, 0x0300, 0x4711, 0},
-    PR29_SUCCESS}
+   /* http://lists.gnu.org/archive/html/help-libidn/2012-01/msg00008.html */
+   "Infloop",
+   3,
+   {0x1100, 0x0300, 0x4711, 0},
+   PR29_SUCCESS}
 };
 
 void
 doit (void)
 {
-  size_t i;
+  unsigned i;
   int rc;
 
   for (i = 0; i < sizeof (tv) / sizeof (tv[0]); i++)
@@ -100,7 +100,7 @@ doit (void)
 	{
 	  uint32_t *p, *q;
 
-	  printf ("PR29 entry %ld: %s\n", i, tv[i].name);
+	  printf ("PR29 entry %u: %s\n", i, tv[i].name);
 
 	  printf ("in:\n");
 	  ucs4print (tv[i].in, tv[i].inlen);
@@ -120,7 +120,7 @@ doit (void)
       rc = pr29_4 (tv[i].in, tv[i].inlen);
       if (rc != tv[i].rc)
 	{
-	  fail ("PR29 entry %ld failed (expected %d): %d\n", i, tv[i].rc, rc);
+	  fail ("PR29 entry %u failed (expected %d): %d\n", i, tv[i].rc, rc);
 	  if (debug)
 	    printf ("FATAL\n");
 	  continue;
@@ -129,7 +129,7 @@ doit (void)
       rc = pr29_4z (tv[i].in);
       if (rc != tv[i].rc)
 	{
-	  fail ("PR29 entry %ld failed (expected %d): %d\n", i, tv[i].rc, rc);
+	  fail ("PR29 entry %u failed (expected %d): %d\n", i, tv[i].rc, rc);
 	  if (debug)
 	    printf ("FATAL\n");
 	  continue;
@@ -142,7 +142,7 @@ doit (void)
 	p = stringprep_ucs4_to_utf8 (tv[i].in, (ssize_t) tv[i].inlen,
 				     &items_read, &items_written);
 	if (p == NULL)
-	  fail ("FAIL: stringprep_ucs4_to_utf8(tv[%ld]) == NULL\n", i);
+	  fail ("FAIL: stringprep_ucs4_to_utf8(tv[%u]) == NULL\n", i);
 	if (debug)
 	  hexprint (p, strlen (p));
 
@@ -150,7 +150,7 @@ doit (void)
 	free (p);
 	if (rc != tv[i].rc)
 	  {
-	    fail ("PR29 entry %ld failed (expected %d): %d\n",
+	    fail ("PR29 entry %u failed (expected %d): %d\n",
 		  i, tv[i].rc, rc);
 	    if (debug)
 	      printf ("FATAL\n");
